pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
aiohttp = "^3.9.1"
aiosqlite = "^0.20.0"
alembic = "^1.13.0"
anthropic = "^0.33.0"
beautifulsoup4 = "^4.12.2"
docker = "^7.0.0"
fire = "^0.6.0"
frozendict = "^2.3.10"
frozenlist = "^1.4.1"
huggingface-hub = "^0.24.5"
jq = "^1.7.0"
jsonschema = "^4.20.0"
numpy = "^2.0.1"
openai = "^1.4.0"
ormar = "^0.20"
pandas = "^2.1.4"
protobuf = "^5.27.3"
pydantic = "^2.0"
pyte = "^0.8.2"
python = "^3.11"
python-dotenv = "^1.0.1"
regex = "^2024.7.24"
requests = "^2.31.0"
toml = "^0.10.2"
tqdm = "^4.66.1"
transformers = "^4.36.1"
pytest = "^8.3.2"
replicate = "^0.32.1"
rich = "^13.9.2"
